Fix JWT authentication
The wrong dispatch function was used first, now it is
implemented in dispatch but also in the _auth function.

Also, tests have been added to make sure that connecting works.
I did not choose to use the requests-jwt library as I could
not make it work, and the few lines of code that are needed
are so slim that another dependency is not warranted.